Move Call/Return kernels and ops to separate file
TODO: Must check that bazel hooks are placed appropriately for the
proper registration of the new ops.
