+++ tensorflow/core/BUILD
@@ -567,6 +567,7 @@ tf_gen_op_libs(
         "data_flow_ops",
         "dataset_ops",
         "function_ops",
+        "function_control_ops",
         "functional_ops",
         "image_ops",
         "io_ops",
@@ -648,6 +649,7 @@ cc_library(
         ":dataset_ops_op_lib",
         ":function_ops_op_lib",
         ":functional_ops_op_lib",
+        ":function_control_ops_op_lib",
         ":image_ops_op_lib",
         ":io_ops_op_lib",
         ":linalg_ops_op_lib",
@@ -780,6 +782,7 @@ cc_library(
         "//tensorflow/core/kernels:dataset_ops",
         "//tensorflow/core/kernels:fake_quant_ops",
         "//tensorflow/core/kernels:function_ops",
+        "//tensorflow/core/kernels:function_control_ops",
         "//tensorflow/core/kernels:image",
         "//tensorflow/core/kernels:io",
         "//tensorflow/core/kernels:linalg",
