+++ tensorflow/core/kernels/BUILD
@@ -1437,6 +1437,16 @@ tf_cc_test(
     ],
 )
 
+tf_kernel_library(
+    name = "function_control_ops",
+    prefix = "function_control_ops",
+    deps = [
+        "//tensorflow/core:function_control_op_lib",
+        "//tensorflow/core:framework",
+        "//tensorflow/core:lib",
+    ],
+)
+
 cc_library(
     name = "data_flow",
     deps = [
